Updated structure of the project to be more templated and added additional pages
