+++ frontend/components/Meta.js
@@ -1,0 +1,13 @@
+import Head from "next/head";
+
+const Meta = () => (
+  <Head>
+    <meta name='viewport' content='width=device-width, initial-scale=1' />
+    <meta charSet='utf-8' />
+    <link rel='shortcut icon' href='/favicon.png' />
+    <link rel='stylesheet' type='text/css' href='/nprogress.css' />
+    <title className='print-title'>Energy Assistance Dashboard</title>
+  </Head>
+);
+
+export default Meta;
