+++ frontend/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "next": "9.4.4",
+    "nprogress": "^0.2.0",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-vis": "^1.11.7",
