+++ frontend/pages/about.js
@@ -1,0 +1,11 @@
+// TODO: Create and style about page
+
+function About(props) {
+  return(
+    <div>
+      About Page
+    </div>
+  )
+}
+
+export default About;
