+++ frontend/pages/counties/[id].js
@@ -1,0 +1,10 @@
+import { useRouter } from 'next/router'
+
+const County = () => {
+  const router = useRouter()
+  const { id } = router.query
+
+  return <p>County: {id}</p>
+}
+
+export default County
